+++ src/main/java/entity/IEntity.java
@@ -4,5 +4,5 @@ package entity;
  * Created by EgorVeremeychik on 18.04.2016.
  */
 public interface IEntity {
-    long getId();
+    Integer getId();
 }
